+++ quantify_scheduler/backends/graph_compilation.py
@@ -20,7 +20,7 @@ from typing import (
 import matplotlib.pyplot as plt
 import networkx as nx
 from matplotlib.axes import Axes
-from pydantic import field_serializer, field_validator
+from pydantic import Field, field_serializer, field_validator
 
 from quantify_scheduler.backends.types.common import HardwareCompilationConfig
 from quantify_scheduler.helpers.importers import export_python_object_to_path_string
@@ -175,7 +175,33 @@ class DeviceCompilationConfig(DataStructure):
     The scheduling strategy used when determining the absolute timing of each
     operation of the schedule.
     """
-    compilation_passes: List[SimpleNodeConfig] = []
+    compilation_passes: List[SimpleNodeConfig] = Field(
+        default=[
+            {
+                "name": "circuit_to_device",
+                "compilation_func": "quantify_scheduler.backends.circuit_to_device."
+                + "_compile_circuit_to_device",
+            },
+            {
+                "name": "set_pulse_and_acquisition_clock",
+                "compilation_func": "quantify_scheduler.backends.circuit_to_device."
+                + "set_pulse_and_acquisition_clock",
+            },
+            {
+                "name": "resolve_control_flow",
+                "compilation_func": "quantify_scheduler.compilation.resolve_control_flow",
+            },
+            {
+                "name": "determine_absolute_timing",
+                "compilation_func": "quantify_scheduler.compilation._determine_absolute_timing",
+            },
+            {
+                "name": "flatten",
+                "compilation_func": "quantify_scheduler.compilation.flatten_schedule",
+            },
+        ],
+        validate_default=True,
+    )
     """
     The list of compilation nodes that should be called in succession to compile a
     schedule to the quantum-device layer.